Cleaning and moving props part 15
